requirements.txt
google-cloud-bigquery==2.*
pandas==1.4.*
pyarrow==8.*
